bitly/requirements.txt
pytest==5.4.2
pytest-html==2.1.1
PyYAML==5.3.1
requests==2.23.0
